dumpDagTest() -> BoundDagTest.GetDebuggerDisplay()
